test: 테스트 코드 클래스 분리 및 개발 기능 테스트 코드 추가
